requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
coremltools==5.2.0
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
imutils==0.5.4
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.4
jupyter-core==4.11.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
kraken
loguru==0.6.0
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
praat-parselmouth==0.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytorch-lightning==1.6.3
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzbar==0.1.9
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
scikit-image==0.19.2
scipy==1.8.1
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.19.1
sympy==1.10.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
tifffile==2022.5.4
tinycss2==1.1.1
torch==1.11.0
torchmetrics==0.9.2
torchvision
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.9
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
yarl==1.7.2
zipp==3.8.0
